locked down users mutations

+++ src/resolvers/Mutation.js
@@ -92,7 +92,7 @@ const Mutation = {
       }
     })
 
-    if(!postExists) {
+    if (!postExists) {
       throw new Error('Unable to delete to post ')
     }
 
@@ -103,7 +103,20 @@ const Mutation = {
     }, info)
   },
 
-  updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!postExists) {
+      throw new Error('Unable to update post ')
+    }
+
     return prisma.mutation.updatePost({
       where: {
         id: args.id
@@ -112,13 +125,15 @@ const Mutation = {
     }, info)
   },
 
-  createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -130,7 +145,20 @@ const Mutation = {
     }, info)
   },
 
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) {
+      throw new Error('Unable to delete comment')
+    }
+
     return prisma.mutation.deleteComment({
       where: {
         id: args.id
@@ -138,7 +166,20 @@ const Mutation = {
     }, info)
   },
 
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) {
+      throw new Error('Unable to update comment')
+    }
+
     return prisma.mutation.updateComment({
       where: {
         id: args.id
